Rollback to RxJava event bus implementation
The vertx event bus implementation was not easily usable, since
I had to register custom codecs for every event we sent through
the bus.
Pojos could be converted to byte arrays which is always sendable
through the bus, but that had 2 problems:it needed the
ObjectMapper as another dependency in a) the aggregate repository
(to convert pojos to byte arrays) and b) also in every event handler
class to convert it back.

I'll go now with the Rx implementation, maybe someday it could be
implemented using kotlin coroutines.